+++ exec/bsf_chipseq_diffbind.R
@@ -86,7 +86,7 @@ argument_list <-
         default = 0.05,
         dest = "fdr_threshold",
         help = "FDR threshold [0.05]",
-        type = "numeric"
+        type = "double"
       ),
       optparse::make_option(
         opt_str = "--threads",
@@ -100,14 +100,14 @@ argument_list <-
         default = 7.0,
         dest = "plot_width",
         help = "Plot width in inches [14.0]",
-        type = "numeric"
+        type = "double"
       ),
       optparse::make_option(
         opt_str = "--plot-height",
         default = 7.0,
         dest = "plot_height",
         help = "Plot height in inches [36.0]",
-        type = "numeric"
+        type = "double"
       )
     )
   ))
