+++ exec/bsf_rnaseq_deseq_analysis.R
@@ -101,7 +101,7 @@ argument_list <-
         default = 0.1,
         dest = "padj_threshold",
         help = "Adjusted p-value threshold [0.1]",
-        type = "numeric"
+        type = "double"
       ),
       optparse::make_option(
         # This option is required for PCA plots
@@ -145,21 +145,21 @@ argument_list <-
       #   default = 0.5,
       #   dest = "plot_factor",
       #   help = "Plot width increase per 24 samples [0.5]",
-      #   type = "numeric"
+      #   type = "double"
       # ),
       optparse::make_option(
         opt_str = "--plot-width",
         default = 7.0,
         dest = "plot_width",
         help = "Plot width in inches [7.0]",
-        type = "numeric"
+        type = "double"
       ),
       optparse::make_option(
         opt_str = "--plot-height",
         default = 7.0,
         dest = "plot_height",
         help = "Plot height in inches [7.0]",
-        type = "numeric"
+        type = "double"
       )
     )
   ))
@@ -1324,8 +1324,8 @@ plot_pca <- function(object,
           "component_2" = factor(levels = paste0(
             "PC", seq_len(length.out = pca_dimensions)
           )),
-          "x" = numeric(),
-          "y" = numeric(),
+          "x" = double(),
+          "y" = double(),
           # Also initialise all variables of the column data, but do not
           # include any data (i.e., 0L rows).
           S4Vectors::as.data.frame(x = SummarizedExperiment::colData(x = object)[0L,])
